keep intermediate files if DEBUG >= 2

+++ pnmcc/fr.lip6.move.gal.application.pnmcc/src/fr/lip6/move/gal/application/runner/ltsmin/LTSminRunner.java
@@ -48,7 +48,7 @@ public class LTSminRunner extends AbstractRunner implements ILTSminRunner {
 		this.onlyGal = onlyGal;
 		try {
 			this.workFolder = Files.createTempDirectory("ltsmin").toFile();
-			workFolder.deleteOnExit();
+			if (DEBUG < 2) workFolder.deleteOnExit();
 		} catch (IOException e) {
 			System.out.println("Unable to create temporary folder.");
 		}
